refactor: todos completed list
