+++ src/components/Todo/Todo.tsx
@@ -1,36 +1,10 @@
 import React, { useState } from "react";
-
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import {
-  Container,
-  makeStyles,
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Typography,
-} from "@material-ui/core";
-import { AnimatePresence, Reorder } from "framer-motion";
-import { Item, TodoCompletedItem } from "./common";
+import { Container } from "@material-ui/core";
+import { Reorder } from "framer-motion";
+import { Item, TodoCompletedList } from "./common";
 import { Form } from "./common/Todo/Form";
 import { TodoItem } from "./common/types";
 
-const useStyles = makeStyles({
-  accordion: {
-    boxShadow: "none",
-    borderTop: "1px solid black",
-    marginTop: "20px",
-  },
-  accordionDetails: {
-    display: "block",
-    paddingLeft: "0px",
-  },
-  accordionSummary: {
-    paddingLeft: "4.2%",
-  },
-  reorderItem: {
-    listStyle: "none",
-  },
-});
 
 export interface TodoAppProps {
   defaultItems?: TodoItem[];
@@ -40,7 +14,7 @@ export interface TodoAppProps {
 function TodoApp(props: TodoAppProps) {
   const { defaultItems = [], onChange } = props;
   const [items, setItems] = useState<TodoItem[]>(defaultItems);
-  const classes = useStyles();
+  
   const setItemsCallback = (updatedItems: TodoItem[]) => {
     setItems(updatedItems);
     onChange(updatedItems);
@@ -95,30 +69,11 @@ function TodoApp(props: TodoAppProps) {
           );
         })}
       </Reorder.Group>
-      {completedItemsLength > 0 && (
-        <Accordion className={classes.accordion} defaultExpanded={true}>
-          <AccordionSummary
-            className={classes.accordionSummary}
-            expandIcon={<ExpandMoreIcon />}
-            aria-controls="panel1a-content"
-            id="panel1a-header"
-          >
-            <Typography> {completedItemsLength} Completed items </Typography>
-          </AccordionSummary>
-          <AccordionDetails className={classes.accordionDetails}>
-            {items.map((item, index) => {
-              return (
-                <TodoCompletedItem
-                  items={items}
-                  key={item.uuid}
-                  itemIndex={index}
-                  setItemsCallback={setItemsCallback}
-                />
-              );
-            })}
-          </AccordionDetails>
-        </Accordion>
-      )}
+      <TodoCompletedList 
+        items={items}
+        setItemsCallback={setItemsCallback}
+        completedItems={completedItems}
+      />
     </Container>
   );
 }
